Unifies MCP implementation into main application backend
